+++ verify/BUILD
@@ -23,7 +23,7 @@ go_library(
     srcs = ["verify.go"],
     importpath = "github.com/project-oak/transparent-release/verify",
     deps = [
-        "//amber",
+        "//pkg/amber",
         "//common",
         "@com_github_in_toto_in_toto_golang//in_toto/slsa_provenance/v0.2:go_default_library",
     ],
